requirements.txt
pytorch_lightning==1.3.3
torch>=1.6.0
segmentation_models_pytorch
imgaug
omegaconf
pandas
scipy
libtiff
pillow
tqdm
mlflow
git+git://github.com/PyTorchLightning/metrics.git